mapper: Change metadata_len parameter to uint64_t

+++ src/mapper-version0.h
@@ -61,7 +61,7 @@ struct xseg_request * prepare_write_object_v0(struct peer_req *pr,
 				struct map *map, struct map_node *mn);
 //int read_map_v0(struct map *m, unsigned char * data);
 int read_map_metadata_v0(struct map *map, unsigned char *metadata,
-		uint32_t metadata_len);
+		uint64_t metadata_len);
 int load_map_data_v0(struct peer_req *pr, struct map *map);
 int write_map_metadata_v0(struct peer_req *pr, struct map *map);
 int write_map_data_v0(struct peer_req *pr, struct map *map);
