+++ src/mapper-versions.h
@@ -103,7 +103,7 @@ struct map_functions {
 //	int (*read_map)(struct map *map, unsigned char * data);
 //	int (*write_map)(struct peer_req *pr, struct map *map);
 	int (*read_map_metadata)(struct map *map, unsigned char *metadata,
-			uint32_t metadata_len);
+			uint64_t metadata_len);
 	int (*load_map_data)(struct peer_req *pr, struct map *map);
 	int (*write_map_metadata)(struct peer_req *pr, struct map *map);
 	int (*write_map_data)(struct peer_req *pr, struct map *map);
